requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
transformers
torch
spacy
scikit-learn
pandas
numpy
httpx
aiofiles
python-multipart
python-dotenv
structlog
beautifulsoup4
requests
nltk
textstat
pytest
pytest-asyncio
black
isort
flake8
mypy